Rename headers so clang-tidy knows they're C++; Fix readability-else-after-return warnings; Move globals out of config.hpp in prep for refactoring
