+++ platformio.ini
@@ -20,7 +20,7 @@ lib_deps =
 	https://github.com/SwiCago/HeatPump
 check_tool = clangtidy, cppcheck
 check_flags =
-  clangtidy: --checks '-*,bugprone-*,clang-analyzer-*,cppcoreguidelines-*,misc-*,modernize-*,performance-*,portability-*,readability-*,google-*'
+  clangtidy: --checks '-*,bugprone-*,clang-analyzer-*,cppcoreguidelines-*,misc-*,performance-*,portability-*,readability-*,google-*,-cppcoreguidelines-avoid-c-arrays'
   cppcheck: --enable=all --inline-suppr --std=c++20 --suppress=*:*/.pio/* --suppress=missingIncludeSystem --error-exitcode=1
 check_skip_packages = yes
 check_src_filters = +<**/*>
