+++ src/mitsubishi2mqtt/html_common.hpp
