+++ src/mitsubishi2mqtt/html_init.hpp
