+++ src/mitsubishi2mqtt/html_metrics.hpp
