+++ src/mitsubishi2mqtt/javascript_common.hpp
